+++ qml/pages/SettingsPage.qml
@@ -18,7 +18,38 @@ Item {
             spacing: 10
 
             SettingsBox {
+                title: "Settings Category #1"
                 Layout.fillWidth: true
+
+                SettingRow {
+                    title: "Setting #1"
+
+                    UiButton {
+                        buttonHeight: 32
+                        buttonWidth: 170
+
+                        fillColor: fillColor
+                        borderColor: borderColor
+
+                        buttonText: "Button #1"
+                        buttonIcon: "qrc:/qml/assets/icons/download.svg"
+                    }
+                }
+
+                SettingRow {
+                    title: "Setting #2"
+
+                    UiButton {
+                        buttonHeight: 32
+                        buttonWidth: 170
+
+                        fillColor: fillColor
+                        borderColor: borderColor
+
+                        buttonText: "Button #2"
+                        buttonIcon: "qrc:/qml/assets/icons/download.svg"
+                    }
+                }
             }
         }
     }